Add Feature: Delete tasks
This latest update does add a basic delete funciton to the application, but it is not complete yet. Delete needs to not only move from the array, but you have to appened back into the original array, then save it to storage. Will work this out in the next push.