Add: UI/UX icons and styling
Added some basic styling and icons for the next features which will include editing of tasks and deleting of a full list. I am holding off on the editing of a task because of the scripting involved - want to transform the task into an input field that has a save button. This will come later.